Lint Report and suggested improvements implemented

+++ src/main/java/acme/features/client/progresslog/ClientProgressLogDeleteService.java
@@ -1,5 +1,5 @@
 
-package acme.features.client.progressLog;
+package acme.features.client.progresslog;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -45,28 +45,36 @@ public class ClientProgressLogDeleteService extends AbstractService<Client, Prog
 		super.getBuffer().addData(object);
 	}
 
+
+	private static String invalidObject = "Invalid object: ";
+
+
 	@Override
 	public void bind(final ProgressLog object) {
-		assert object != null;
+		if (object == null)
+			throw new IllegalArgumentException(ClientProgressLogDeleteService.invalidObject + object);
 
 		super.bind(object, "recordId", "completeness", "comment", "registrationMoment", "responsiblePerson");
 	}
 
 	@Override
 	public void validate(final ProgressLog object) {
-		assert object != null;
+		if (object == null)
+			throw new IllegalArgumentException(ClientProgressLogDeleteService.invalidObject + object);
 	}
 
 	@Override
 	public void perform(final ProgressLog object) {
-		assert object != null;
+		if (object == null)
+			throw new IllegalArgumentException(ClientProgressLogDeleteService.invalidObject + object);
 
 		this.repository.delete(object);
 	}
 
 	@Override
 	public void unbind(final ProgressLog object) {
-		assert object != null;
+		if (object == null)
+			throw new IllegalArgumentException(ClientProgressLogDeleteService.invalidObject + object);
 
 		Dataset dataset;
